feat: Adding the possibility of copying the link by clicking text and icons (#171)
